+++ src/components/Input.react.js
@@ -118,19 +118,19 @@ Input.propTypes = {
     /**
      * This attribute indicates whether the value of the control can be automatically completed by the browser.
      */
-    autocomplete: PropTypes.string,
+    autoComplete: PropTypes.string,
 
     /**
      * The element should be automatically focused after the page loaded.
      */
-    autofocus: PropTypes.string,
+    autoFocus: PropTypes.string,
 
     /**
      * If true, the input is disabled and can't be clicked on.
      */
     disabled: PropTypes.bool,
 
-    inputmode: PropTypes.oneOf([
+    inputMode: PropTypes.oneOf([
         /**
          * Alphanumeric, non-prose content such as usernames and passwords.
          */
@@ -205,7 +205,7 @@ Input.propTypes = {
     /**
      * If the value of the type attribute is text, email, search, password, tel, or url, this attribute specifies the maximum number of characters (in UTF-16 code units) that the user can enter. For other control types, it is ignored. It can exceed the value of the size attribute. If it is not specified, the user can enter an unlimited number of characters. Specifying a negative number results in the default behavior (i.e. the user can enter an unlimited number of characters). The constraint is evaluated only when the value of the attribute has been changed.
      */
-    maxlength: PropTypes.string,
+    maxLength: PropTypes.string,
 
     /**
      * The minimum (numeric or date-time) value for this item, which must not be greater than its maximum (max attribute) value.
@@ -215,7 +215,7 @@ Input.propTypes = {
     /**
      * If the value of the type attribute is text, email, search, password, tel, or url, this attribute specifies the minimum number of characters (in Unicode code points) that the user can enter. For other control types, it is ignored.
      */
-    minlength: PropTypes.string,
+    minLength: PropTypes.string,
 
     /**
      * This Boolean attribute indicates whether the user can enter more than one value. This attribute applies when the type attribute is set to email or file, otherwise it is ignored.